+++ src/main/java/tavebalak/OTTify/community/dto/request/CommunitySubjectEditDTO.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.community.dto;
+package tavebalak.OTTify.community.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
