+++ src/main/java/tavebalak/OTTify/community/dto/request/ReplyRecommentEditDTO.java
@@ -1,23 +1,22 @@
-package tavebalak.OTTify.community.dto;
+package tavebalak.OTTify.community.dto.request;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.validation.constraints.NotBlank;
-
 @Getter
 @NoArgsConstructor
-public class ReplyRecommentCreateDTO {
+public class ReplyRecommentEditDTO {
     private Long subjectId;
     private Long commentId;
-    @NotBlank(message = "대댓글 내용이 비워져 있어서는 안됩니다.")
+    private Long recommentId;
     private String content;
 
     @Builder
-    public ReplyRecommentCreateDTO(Long subjectId, Long commentId, String content) {
+    public ReplyRecommentEditDTO(Long subjectId, Long commentId, Long recommentId, String content) {
         this.subjectId = subjectId;
         this.commentId = commentId;
+        this.recommentId = recommentId;
         this.content = content;
     }
 }
