+++ src/main/java/tavebalak/OTTify/community/dto/response/CommunityAriclesDTO.java
@@ -1,11 +1,9 @@
-package tavebalak.OTTify.community.dto;
+package tavebalak.OTTify.community.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import tavebalak.OTTify.community.entity.Reply;
 
-import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
