+++ src/main/java/tavebalak/OTTify/community/dto/response/CommunitySubjectsListDTO.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.community.dto;
+package tavebalak.OTTify.community.dto.response;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
