+++ src/main/java/tavebalak/OTTify/community/dto/response/ReplyCommentEditorDTO.java
@@ -1,6 +1,5 @@
-package tavebalak.OTTify.community.dto;
+package tavebalak.OTTify.community.dto.response;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
