+++ src/main/java/tavebalak/OTTify/community/service/CommunityService.java
@@ -2,11 +2,13 @@ package tavebalak.OTTify.community.service;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.multipart.MultipartFile;
-import tavebalak.OTTify.community.dto.*;
+import tavebalak.OTTify.community.dto.request.CommunitySubjectCreateDTO;
+import tavebalak.OTTify.community.dto.request.CommunitySubjectEditDTO;
+import tavebalak.OTTify.community.dto.response.CommunityAriclesDTO;
+import tavebalak.OTTify.community.dto.response.CommunitySubjectDTO;
+import tavebalak.OTTify.community.dto.response.CommunitySubjectsDTO;
 import tavebalak.OTTify.community.entity.Community;
 import tavebalak.OTTify.exception.NotFoundException;
-
-import java.util.List;
 
 public interface CommunityService {
     Community saveSubject(MultipartFile image, CommunitySubjectCreateDTO c);
