+++ src/main/java/tavebalak/OTTify/community/service/ReplyServiceImpl.java
@@ -4,7 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
-import tavebalak.OTTify.community.dto.*;
+import tavebalak.OTTify.community.dto.request.ReplyCommentCreateDTO;
+import tavebalak.OTTify.community.dto.request.ReplyCommentEditDTO;
+import tavebalak.OTTify.community.dto.request.ReplyRecommentCreateDTO;
+import tavebalak.OTTify.community.dto.request.ReplyRecommentEditDTO;
+import tavebalak.OTTify.community.dto.response.CommentDTO;
+import tavebalak.OTTify.community.dto.response.ReplyCommentEditorDTO;
 import tavebalak.OTTify.community.entity.Community;
 import tavebalak.OTTify.community.entity.Reply;
 import tavebalak.OTTify.community.repository.CommunityRepository;
