+++ src/main/java/tavebalak/OTTify/program/controller/ProgramMainController.java
@@ -2,11 +2,10 @@ package tavebalak.OTTify.program.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import tavebalak.OTTify.common.BaseResponse;
-import tavebalak.OTTify.program.dto.RecommendProgramsDTO;
+import tavebalak.OTTify.program.dto.response.RecommendProgramsDTO;
 import tavebalak.OTTify.program.service.ProgramService;
 
 @RestController
@@ -16,7 +15,7 @@ public class ProgramMainController {
     private final ProgramService programService;
 
     @GetMapping("/recommendProgram")
-    public BaseResponse getRecommendProgram(){
+    public BaseResponse<RecommendProgramsDTO> getRecommendProgram(){
         RecommendProgramsDTO recommendProgramsDTO = programService.getRecommendProgram(6);
         return BaseResponse.success(recommendProgramsDTO);
     }
