+++ src/main/java/tavebalak/OTTify/program/service/ProgramService.java
@@ -1,6 +1,6 @@
 package tavebalak.OTTify.program.service;
 
-import tavebalak.OTTify.program.dto.RecommendProgramsDTO;
+import tavebalak.OTTify.program.dto.response.RecommendProgramsDTO;
 
 public interface ProgramService {
     public RecommendProgramsDTO getRecommendProgram(int count);
