+++ src/main/java/tavebalak/OTTify/program/service/ProgramServiceImpl.java
@@ -13,11 +13,10 @@ import tavebalak.OTTify.genre.repository.GenreRepository;
 import tavebalak.OTTify.genre.repository.ProgramGenreRepository;
 import tavebalak.OTTify.genre.repository.UserGenreRepository;
 import tavebalak.OTTify.oauth.jwt.SecurityUtil;
-import tavebalak.OTTify.program.dto.RecommendProgramsDTO;
-import tavebalak.OTTify.program.dto.ServiceListsDTO;
+import tavebalak.OTTify.program.dto.response.RecommendProgramsDTO;
+import tavebalak.OTTify.program.dto.response.ServiceListsDTO;
 import tavebalak.OTTify.program.entity.Program;
 import tavebalak.OTTify.program.repository.ProgramRepository;
-import tavebalak.OTTify.user.entity.LikedProgram;
 import tavebalak.OTTify.user.entity.User;
 import tavebalak.OTTify.user.repository.LikedProgramRepository;
 import tavebalak.OTTify.user.repository.UserRepository;
