+++ src/test/java/tavebalak/OTTify/community/controller/CommunityControllerTest.java
@@ -13,10 +13,9 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import tavebalak.OTTify.community.dto.*;
+import tavebalak.OTTify.community.dto.request.*;
 import tavebalak.OTTify.community.entity.Community;
 import tavebalak.OTTify.community.entity.Reply;
-import tavebalak.OTTify.community.service.CommunityService;
 import tavebalak.OTTify.community.service.CommunityServiceImpl;
 import tavebalak.OTTify.community.service.ReplyService;
 import tavebalak.OTTify.exception.NotFoundException;
